LibraryBaseElement: Allow saving to empty destination
Unfortunately empty directories are sometimes not removed properly, for
example some Git operations remove the contained files but not the
parent directories. But if the destination directory exists already,
FileUtils::copyDirRecursively() would throw an exception because it
requires that the destination does not exist yet. To avoid this issue,
we remove the destination first if it's an empty directory. If it's not
empty, we accept that an exception is thrown because this would likely
be a serious error.